# gen/openapiv2/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "lilbattle/v1/models/indexer.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/games": {
      "get": {
        "summary": "ListGames returns all available games",
        "operationId": "GamesService_ListGames",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListGamesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GamesService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new game",
        "operationId": "GamesService_CreateGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateGameRequest"
            }
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
...
    "/v1/games/{gameId}": {
      "patch": {
        "summary": "GetGame returns a specific game with metadata",
        "operationId": "GamesService_UpdateGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "gameId",
            "description": "Game id to modify",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "newGame": {
                  "$ref": "#/definitions/v1Game",
                  "title": "*\nGame being updated"
                },
                "newState": {
                  "$ref": "#/definitions/v1GameState",
                  "title": "New world state to save"
                },
                "newHistory": {
                  "$ref": "#/definitions/v1GameMoveHistory",
                  "title": "History to save"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this Game to make partial changes."
                }
              },
              "title": "UpdateGameRequest"
            }
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
...
    "/v1/games/{id}": {
      "get": {
        "summary": "GetGame returns a specific game with metadata",
        "operationId": "GamesService_GetGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GamesService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular game",
        "operationId": "GamesService_DeleteGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the game to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
    "/v1/games:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple games by ID",
        "operationId": "GamesService_GetGames",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGamesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the game to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com. As of May 2023, there are no widely used type server\nimplementations and no plans to implement one.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateGameRequest": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being updated"
        }
      },
      "title": "*\nGame creation request object"
    },
    "v1CreateGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being created"
        },
        "gameState": {
          "$ref": "#/definitions/v1GameState",
          "title": "The starting game state"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an game creation."
    },
...
    "v1DeleteGameResponse": {
      "type": "object",
      "title": "*\nGame deletion response"
    },
...
    "v1Game": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "version": {
          "type": "string",
          "format": "int64",
          "title": "Version number for optimistic locking"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the game"
        },
        "creatorId": {
          "type": "string",
          "title": "User who started/created the game"
        },
        "worldId": {
          "type": "string",
          "title": "The world this game was created from"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if game has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - example attribute"
        },
        "config": {
          "$ref": "#/definitions/v1GameConfiguration",
          "title": "Game configuration"
        },
        "previewUrls": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "URL to screenshot/preview image (defaults to /games/{id}/screenshots/{screenshotName})\nCan be overridden to point to CDN or external hosting"
        },
        "searchIndexInfo": {
          "$ref": "#/definitions/v1IndexInfo"
        }
      },
      "title": "Describes a game and its metadata"
    },
    "v1GameConfiguration": {
      "type": "object",
      "properties": {
        "players": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GamePlayer"
          },
          "title": "Player configuration"
        },
        "teams": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GameTeam"
          },
          "title": "Team configuration"
        },
        "incomeConfigs": {
          "$ref": "#/definitions/v1IncomeConfig",
          "title": "Various kinds of per turn income configs"
        },
        "settings": {
          "$ref": "#/definitions/v1GameSettings",
          "title": "Game settings"
        }
      }
    },
...
    "v1GameMoveGroup": {
      "type": "object",
      "properties": {
        "startedAt": {
          "type": "string",
          "format": "date-time",
          "title": "When the moves happened (or were submitted)"
        },
        "endedAt": {
          "type": "string",
          "format": "date-time"
        },
        "groupNumber": {
          "type": "string",
          "format": "int64",
          "title": "Group number within the game - will be monotonically increasing"
        },
        "moves": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GameMove"
          },
          "title": "*\nList of moves to add -"
        }
      },
      "title": "A move group - we can allow X moves in one \"tick\""
    },
    "v1GameMoveHistory": {
      "type": "object",
      "properties": {
        "gameId": {
          "type": "string",
          "title": "Move history for the game"
        },
        "groups": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GameMoveGroup"
          },
          "title": "Each entry in our history is a \"group\" of moves"
        }
      },
      "title": "Holds the game's move history (can be used as a replay log)"
    },
...
    "v1GamePlayer": {
      "type": "object",
      "properties": {
        "playerId": {
          "type": "integer",
          "format": "int32",
          "title": "Player ID (1-based)"
        },
        "userId": {
          "type": "string",
          "title": "ID of the system user that is assigned to this game player.  This is the \"auth\" user"
        },
        "playerType": {
          "type": "string",
          "description": "\"human\", \"ai\", \"open\"",
          "title": "Player type"
        },
        "color": {
          "type": "string",
          "title": "Player color"
        },
        "teamId": {
          "type": "integer",
          "format": "int32",
          "title": "Team ID (0 = no team, 1+ = team number)"
        },
        "name": {
          "type": "string",
          "title": "Nickname for the player in this game"
        },
        "isActive": {
          "type": "boolean",
          "title": "Whether play is still in the game - can this just be inferred?"
        },
        "startingCoins": {
          "type": "integer",
          "format": "int32",
          "title": "How many coins the player started off with"
        }
      }
    },
...
    "v1GameState": {
      "type": "object",
      "properties": {
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "gameId": {
          "type": "string",
          "title": "ID of the game whos state is being tracked"
        },
        "turnCounter": {
          "type": "integer",
          "format": "int32"
        },
        "currentPlayer": {
          "type": "integer",
          "format": "int32"
        },
        "worldData": {
          "$ref": "#/definitions/v1WorldData",
          "title": "Current world state"
        },
        "stateHash": {
          "type": "string",
          "title": "Current state hash for validation"
        },
        "version": {
          "type": "string",
          "format": "int64",
          "title": "Version number for optimistic locking"
        },
        "status": {
          "$ref": "#/definitions/v1GameStatus"
        },
        "finished": {
          "type": "boolean",
          "title": "Only set after a win has been possible"
        },
        "winningPlayer": {
          "type": "integer",
          "format": "int32"
        },
        "winningTeam": {
          "type": "integer",
          "format": "int32"
        },
        "currentGroupNumber": {
          "type": "string",
          "format": "int64"
        },
        "playerStates": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1PlayerState"
          },
          "title": "Per-player runtime state, keyed by player_id (1-based)\nThis holds mutable player state like coins that changes during gameplay"
        }
      },
      "title": "Holds the game's Active/Current state (eg world state)"
    },
    "v1GameStatus": {
      "type": "string",
      "enum": [
        "GAME_STATUS_UNSPECIFIED",
        "GAME_STATUS_PLAYING",
        "GAME_STATUS_PAUSED",
        "GAME_STATUS_ENDED"
      ],
      "default": "GAME_STATUS_UNSPECIFIED",
      "title": "/////// Game related models"
    },
    "v1GameTeam": {
      "type": "object",
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int32",
          "title": "ID of the team within the game (unique to the game)"
        },
        "name": {
          "type": "string",
          "title": "Name of the team - in a game"
        },
        "color": {
          "type": "string",
          "title": "Just a color for this team"
        },
        "isActive": {
          "type": "boolean",
          "title": "Whether team has active players - can also be inferred"
        }
      }
    },
...
    "v1GetGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game"
        },
        "state": {
          "$ref": "#/definitions/v1GameState"
        },
        "history": {
          "$ref": "#/definitions/v1GameMoveHistory"
        }
      }
    },
...
    "v1GetGamesResponse": {
      "type": "object",
      "properties": {
        "games": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Game"
          }
        }
      },
      "title": "*\nGame batch-get response"
    },
...
    "v1IncomeConfig": {
      "type": "object",
      "properties": {
        "startingCoins": {
          "type": "integer",
          "format": "int32",
          "title": "How much starting coins to give each player at the start of the agme"
        },
        "gameIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income each player just for being in the game each turn"
        },
        "landbaseIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income from each landbase per turn"
        },
        "navalbaseIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income from each navalbase per turn"
        },
        "airportbaseIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income from each airport base per turn"
        },
        "missilesiloIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income from each missile silo per turn"
        },
        "minesIncome": {
          "type": "integer",
          "format": "int32",
          "title": "Income from each mine per turn"
        }
      }
    },
    "v1IndexInfo": {
      "type": "object",
      "properties": {
        "lastUpdatedAt": {
          "type": "string",
          "format": "date-time",
          "title": "We maintain an IndexInfo for each type of \"indexing\" operation needed\nFor example one update may change the keywords (so we need to update indexes for search)\nAnother might update the \"units\" so we may need a new screenshot\nEach one's updates - updated and indexed timestamps separately so they can be tracked sepately"
        },
        "lastIndexedAt": {
          "type": "string",
          "format": "date-time"
        },
        "needsIndexing": {
          "type": "boolean"
        }
      }
    },
...
    "v1ListGamesResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Game"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
...
    "v1PaginationResponse": {
      "type": "object",
      "properties": {
        "nextPageKey": {
          "type": "string",
          "description": "*\nThe key/pointer string that subsequent List requests should pass to\ncontinue the pagination."
        },
        "nextPageOffset": {
          "type": "integer",
          "format": "int32",
          "title": "*\nAlso support an integer offset if possible"
        },
        "hasMore": {
          "type": "boolean",
          "description": "*\nWhether theere are more results."
        },
        "totalResults": {
          "type": "integer",
          "format": "int32",
          "description": "*\nTotal number of results."
        }
      }
    },
...
    "v1PlayerState": {
      "type": "object",
      "properties": {
        "coins": {
          "type": "integer",
          "format": "int32",
          "title": "Current coin balance (changes during gameplay via building, income, etc.)"
        },
        "isActive": {
          "type": "boolean",
          "title": "Whether player is still active in the game (not eliminated)"
        }
      },
      "title": "Runtime state for a player during the game\nThis is separate from GamePlayer (which is player configuration)\nPlayerState is indexed by player_id in the player_states map"
    },
...
    "v1UpdateGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an Game.",
      "title": "UpdateGameResponse"
    },
...
    "v1WorldData": {
      "type": "object",
      "properties": {
        "tilesMap": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Tile"
          },
          "title": "New map-based storage (key = \"q,r\" coordinate string)"
        },
        "unitsMap": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Unit"
          }
        },
        "screenshotIndexInfo": {
          "$ref": "#/definitions/v1IndexInfo",
          "title": "When this world data was updated (may have happened without world updating)"
        },
        "contentHash": {
          "type": "string",
          "title": "We will only update if hash's are different"
        },
        "version": {
          "type": "string",
          "format": "int64",
          "title": "Version for Optimistic concurrent locking"
        },
        "crossings": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Crossing"
          },
          "description": "key = \"q,r\", value = crossing with connectivity",
          "title": "Improvement layer - crossings (roads on land, bridges on water)"
        }
      }
    }
  }
}